chore: remove runInBand option and split config from other tests
