+++ package.json
@@ -16,7 +16,7 @@
   "license": "MIT",
   "main": "src/index.js",
   "scripts": {
-    "test": "jest -i --colors --verbose",
+    "test": "jest --colors --verbose",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "fmt": "prettier --write ."
